Add instances configuration
This will be read by deploy-quizzes.py to manage deployments in the
openSUSE infrastructure. New entries will trigger the creation of an
instance, removed entries will trigger the removal of an instance. The
instances are listed as objects instead of array elements to allow for
future addition of instance specific configuration options, such as
the enabled question catalogues.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>